Moving the protobuf message descriptions into Common
